Added script_raised_build events
Build blueprint provides scirpt_raised_build event with created_entity. Construction order does so, if it is not for a blueprint and only after it has done everything it needed to do to the ghost. I plan to use this with an instant blueprint mod. If the ghost would get revived before the code of the construction order finishes, bad things might happen.